Fix #9 Use case diagram
Define the use case diagram with palette.  Requires seeding
a model of UML Light DI element-types to define a “class as
subject” visual type for Class specifically, to avoid asking the
user to choose from amongst many inappropriate classifier
metaclasses to instantiate.

Also necessary is an assistant model to define a diagram
assistant for the class-as-subject visualization of a class.

Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>