Fix #8 Sequence Diagram and Palette
Implement palette and modeling assistants for light sequence diagram.
The latter requires additionally installing the popup-bar edit-policy
on the interaction compartment for the Lifeline assistant.

Incidentally fix up the “Comment Link” icon in the palette of all
UML Light diagrams to match the class diagram (instead of showing
a comment node, which is misleading).

Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>