Fix #15 Finish up diagram styles
Adds diagram style for activity and sequence diagram, as well as several
improvements including the coloring of associations as discussed in #15,
and refactorings to ease maintenance. Also introduces more consistency,
e.g. element icons of association classes are now hidden. Element icons
are only shown if a similar shape is used for different types of
elements on the same diagram.