Fix #71 Adopt 2018-12 release in the build
- rework the ‘photon’ release TP as 2018-12 and update the
build to use the release target.
- include the equinox transforms dependencies in the release
target platform definition.
-a lso bring the nightly target up to speed with the release target.
- in the 2018-12 release, Eclipse Project removed the SWT
implementation for Windows 32-bit platform.  Only 64-bit
is supported.  This resulted in failure to resolve any SWT
APIs in the Tycho compiler because it uses the first listed
target platform specification, which in the POM happened
to be that now-unsupported 32-bit Windows platform.
So, remove the 32-bit Windows platform from the POM.

Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>