Fix #81 Icon attribution and branding
Redesign the UML Light logo and derived icons using
an original drawing of a feather.  This does not require
attribution as it is licensed to Eclipse.org and others
in the same fashion as other source code contributions.

Fix other branding issues discovered in testing (such as
missing/incorrect bundle names and vendors, feature icon,
descriptions of content, etc. in the About dialog).

Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>